server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pyjwt==2.10.1
pydantic==2.11.1
